requirements.txt
Django==4.2.10
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
